Reorganize component files, rename todos to tasks.service, add property to AssistantContainer
